+++ views/panel/profile.ejs
@@ -82,7 +82,7 @@
   </div>
 
   <script src="https://kit.fontawesome.com/7a20fd358e.js" crossorigin="anonymous"></script>
-  <script type="module" src="/js/tempController.js"></script>
+  <script type="module" src="/js/controller.js"></script>
 </body>
 
 </html>
